# server/docs/swagger.yaml
definitions:
  request.HabitCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      userId:
        type: string
    type: object
  request.HabitUpdateRequest:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  request.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  request.UserUpdateRequest:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - password
    - username
    type: object
  response.HabitResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      userId:
        type: string
    type: object
  response.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: emirhan1usta@gmail.com
    name: emirhan usta
  description: This is a sample swagger for Habit Tracker Fiber REST API
  title: Habit Tracker Fiber REST API
  version: "1.0"
paths:
  /api/v1/habit:
    post:
      consumes:
      - application/json
      description: save habit
      parameters:
      - description: habit
        in: body
        name: habit
        required: true
        schema:
          $ref: '#/definitions/request.HabitCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method save habit
      tags:
      - Habit
    put:
      consumes:
      - application/json
      description: update habit
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/request.HabitUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method update habit
      tags:
      - Habit
  /api/v1/habit-tracker/user/email:
    get:
      consumes:
      - application/json
      description: get user by email
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get user by given email
      tags:
      - User
  /api/v1/habit/{id}:
    delete:
      consumes:
      - application/json
      description: delete habit by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method delete habit by id
      tags:
      - Habit
    get:
      consumes:
      - application/json
      description: get habit by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HabitResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get habit by given id
      tags:
      - Habit
  /api/v1/habit/user/{userId}:
    get:
      consumes:
      - application/json
      description: get all habits by given user id
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.HabitResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get all habits by given user id
      tags:
      - Habit
  /api/v1/user:
    get:
      consumes:
      - application/json
      description: get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.UserResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: saving new user
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        schema:
          $ref: '#/definitions/request.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for saving new user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update user
      tags:
      - User
  /api/v1/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete user by id
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: get user by id
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get user by given id
      tags:
      - User
swagger: "2.0"